fix segfault when no clock is available
When there is no clock available, we even register the display ops.
The 'enter' callbakc is set but not data is available int the row
private data which leads to a segfaults.

This patch fix this problem by not registering the ops if the pm
subsystem was not correctly initialized.

In the meantime, we have to erase the window when we are switching
from one window to another. We can say it is the "default" display
callback.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Reported-by: Milo (Woogyom) Kim <woogyom.kim@gmail.com>