Remove sensor unit
The value exported could be in Degree or Volt. There is no way
to determine the unit used expect to refer the board semantics.

Until the kernel exports the unit associated with this value,
we can not take a decision from the userspace, the value will
be a raw value.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>